Change data type from int to float

+++ database/migrations/2021_05_15_040502_add_order_to_permissions_table.php
@@ -14,7 +14,7 @@ class AddOrderToPermissionsTable extends Migration
     public function up()
     {
         Schema::table('permissions', function (Blueprint $table) {
-            $table->smallInteger('order')->nullable();
+            $table->float('order', 8, 2)->nullable();
         });
     }
